refactor(home): extract tile form-value helpers into their own file
coerceFormValues/seedFormValues/toolInputSummary were pure functions
buried in the 1171-line add-tile-drawer.tsx with no test coverage.
Moved them verbatim to tile-form-values.ts with a unit test.
