+++ apps/mesh/src/web/components/home/add-tile-drawer.tsx
@@ -46,6 +46,11 @@ import {
 import { AgentPromptList } from "./agent-prompt-list";
 import { NativeTilesSection } from "./native-tiles-section";
 import { SectionHeader } from "./section-header";
+import {
+  coerceFormValues,
+  seedFormValues,
+  toolInputSummary,
+} from "./tile-form-values";
 import { Button } from "@deco/ui/components/button.tsx";
 import { Input } from "@deco/ui/components/input.tsx";
 import { ScrollArea } from "@deco/ui/components/scroll-area.tsx";
@@ -731,77 +736,6 @@ function AgentToolList({
       </div>
     </div>
   );
-}
-
-/** Coerce form values for persistence: JSON-parse object/array fields,
- * convert number/integer strings. Returns null on parse error (caller
- * should toast). */
-function coerceFormValues(
-  values: Record<string, unknown>,
-  properties: Record<string, ToolInputProperty>,
-): Record<string, unknown> | null {
-  const out: Record<string, unknown> = {};
-  for (const [key, raw] of Object.entries(values)) {
-    const type = properties[key]?.type;
-    if ((type === "object" || type === "array") && typeof raw === "string") {
-      if (!raw.trim()) continue;
-      try {
-        out[key] = JSON.parse(raw);
-      } catch {
-        return null;
-      }
-    } else if (
-      (type === "number" || type === "integer") &&
-      typeof raw === "string"
-    ) {
-      if (!raw.trim()) continue;
-      out[key] = Number(raw);
-    } else if (raw !== "" && raw != null) {
-      out[key] = raw;
-    }
-  }
-  return out;
-}
-
-/** Stringify toolInput values for form display: object/array values become
- * JSON strings so they render in a Textarea. */
-function seedFormValues(
-  toolInput: Record<string, unknown> | undefined,
-  properties: Record<string, ToolInputProperty> | undefined,
-): Record<string, unknown> {
-  if (!toolInput || !properties) return {};
-  const init: Record<string, unknown> = {};
-  for (const [key, val] of Object.entries(toolInput)) {
-    const type = properties[key]?.type;
-    if (
-      (type === "object" || type === "array") &&
-      val != null &&
-      typeof val !== "string"
-    ) {
-      init[key] = JSON.stringify(val, null, 2);
-    } else {
-      init[key] = val;
-    }
-  }
-  return init;
-}
-
-/** Summarize toolInput for display in the pinned tile row. */
-function toolInputSummary(
-  toolInput: Record<string, unknown> | undefined,
-): string {
-  if (!toolInput) return "";
-  const entries = Object.entries(toolInput).filter(
-    ([, v]) => v != null && v !== "",
-  );
-  if (entries.length === 0) return "";
-  return entries
-    .slice(0, 3)
-    .map(([k, v]) => {
-      const s = typeof v === "object" ? JSON.stringify(v) : String(v);
-      return `${k}=${s.length > 20 ? `${s.slice(0, 20)}…` : s}`;
-    })
-    .join(", ");
 }
 
 /** Row for a pinned tile instance — shows a summary, gear to edit, minus to remove. */
