+++ apps/mesh/src/web/components/home/tile-form-values.test.ts
@@ -1,0 +1,70 @@
+import { describe, expect, test } from "bun:test";
+import {
+  coerceFormValues,
+  seedFormValues,
+  toolInputSummary,
+} from "./tile-form-values";
+
+describe("coerceFormValues", () => {
+  test("parses object/array JSON strings", () => {
+    expect(
+      coerceFormValues({ foo: '{"a":1}' }, { foo: { type: "object" } }),
+    ).toEqual({ foo: { a: 1 } });
+  });
+
+  test("returns null on invalid JSON", () => {
+    expect(
+      coerceFormValues({ foo: "{not json" }, { foo: { type: "object" } }),
+    ).toBeNull();
+  });
+
+  test("converts number/integer strings", () => {
+    expect(coerceFormValues({ n: "42" }, { n: { type: "integer" } })).toEqual({
+      n: 42,
+    });
+  });
+
+  test("drops empty values", () => {
+    expect(coerceFormValues({ s: "" }, { s: { type: "string" } })).toEqual({});
+  });
+});
+
+describe("seedFormValues", () => {
+  test("stringifies object/array values for display", () => {
+    expect(
+      seedFormValues({ foo: { a: 1 } }, { foo: { type: "object" } }),
+    ).toEqual({ foo: JSON.stringify({ a: 1 }, null, 2) });
+  });
+
+  test("passes through non-object values", () => {
+    expect(seedFormValues({ s: "hi" }, { s: { type: "string" } })).toEqual({
+      s: "hi",
+    });
+  });
+
+  test("returns empty object when toolInput or properties missing", () => {
+    expect(seedFormValues(undefined, {})).toEqual({});
+    expect(seedFormValues({ s: "hi" }, undefined)).toEqual({});
+  });
+});
+
+describe("toolInputSummary", () => {
+  test("returns empty string when no toolInput", () => {
+    expect(toolInputSummary(undefined)).toBe("");
+  });
+
+  test("filters empty entries and truncates long values", () => {
+    expect(
+      toolInputSummary({
+        a: "short",
+        b: "",
+        c: null,
+        d: "a".repeat(30),
+      }),
+    ).toBe(`a=short, d=${"a".repeat(20)}…`);
+  });
+
+  test("caps at 3 entries", () => {
+    expect(toolInputSummary({ a: 1, b: 2, c: 3, d: 4 })).toBe("a=1, b=2, c=3");
+  });
+});
